Add files via upload
This is a Java program that calculates the number of 10 rupee coins and 5 rupee coins required to make up a given amount. Here’s how it works:

The program prompts the user to enter the number of test cases.
For each test case, the program prompts the user to enter the amount they have.
If the amount is divisible by 10, the program calculates the number of 10 rupee coins required to make up the amount and prints the result.
If the amount is not divisible by 10, the program checks if the remainder when the amount is divided by 10 is divisible by 5 and if the amount is greater than 5 and less than 10. If both conditions are true, the program calculates the number of 10 rupee coins and 5 rupee coins required to make up the amount and prints the result. Otherwise, the program prints “impossible”.
I hope that helps! Let me know if you have any more questions.